requirements.txt
numpy~=1.18.5
torch
gdown~=4.4.0
natsort~=8.1.0
pillow~=9.0.0
torchvision~=0.11.1
tqdm~=4.47.0
requests~=2.24.0

matplotlib~=3.2.2